pyproject.toml
[project]
name = "vera"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiofiles>=25.1.0",
    "aiosqlite>=0.22.1",
    "anthropic>=0.86.0",
    "blinker>=1.9.0",
    "casbin>=1.43.0",
    "chromadb>=1.5.5",
    "cryptography>=46.0.5",
    "fastapi>=0.135.2",
    "keyring>=25.7.0",
    "ollama>=0.6.1",
    "openai>=2.29.0",
    "opentelemetry-api>=1.40.0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.40.0",
    "opentelemetry-exporter-otlp>=1.40.0",
    "opentelemetry-sdk>=1.40.0",
    "presidio-analyzer>=2.2.362",
    "presidio-anonymizer>=2.2.362",
    "prometheus-client>=0.24.1",
    "pydantic-ai>=0.8.1",
    "pydantic>=2.12.5",
    "pyyaml>=6.0.3",
    "rocketry>=2.5.1",
    "spacy>=3.8.13",
    "sqlalchemy[asyncio]>=2.0.48",
    "streamlit>=1.55.0",
    "typer>=0.24.1",
    "uvicorn[standard]>=0.42.0",
]


[tool.setuptools.packages.find]
include = ["core*", "interfaces*", "plugins*"]

[project.scripts]
vera = "interfaces.cli.main:app"

[tool.pytest.ini_options]
pythonpath = "."
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]

[tool.pyright]
venvPath = "."
venv = ".venv"
